C++: use `from_` prefix for static constructors (#3990)
### What
* Closes https://github.com/rerun-io/rerun/issues/3957

### Checklist
* [x] I have read and agree to [Contributor
Guide](https://github.com/rerun-io/rerun/blob/main/CONTRIBUTING.md) and
the [Code of
Conduct](https://github.com/rerun-io/rerun/blob/main/CODE_OF_CONDUCT.md)
* [x] I've included a screenshot or gif (if applicable)
* [x] I have tested [demo.rerun.io](https://demo.rerun.io/pr/3990) (if
applicable)
* [x] The PR title and labels are set such as to maximize their
usefulness for the next release's CHANGELOG

- [PR Build Summary](https://build.rerun.io/pr/3990)
- [Docs
preview](https://rerun.io/preview/644af63af4e0dd6b1941cc03ffd1142f00429cfa/docs)
<!--DOCS-PREVIEW-->
- [Examples
preview](https://rerun.io/preview/644af63af4e0dd6b1941cc03ffd1142f00429cfa/examples)
<!--EXAMPLES-PREVIEW-->
- [Recent benchmark results](https://ref.rerun.io/dev/bench/)
- [Wasm size tracking](https://ref.rerun.io/dev/sizes/)

---------

+++ rerun_cpp/src/rerun/demo_utils.cpp
@@ -20,7 +20,9 @@ namespace rerun {
                     static_cast<float>(i) * z_step,
                 });
 
-                out_colors.push_back(colormap_turbo_srgb(static_cast<float>(i) / num_points));
+                out_colors.push_back(
+                    colormap_turbo_srgb(static_cast<float>(i) / static_cast<float>(num_points))
+                );
             }
         }
